+++ examples/vue2/src/demos/infinite-scroll.vue
@@ -92,7 +92,7 @@ onMounted(() => {
     bind(containerRef.value)
   }
 })
-</script>
+<\/script>
 
 <template>
   <div ref="containerRef" class="scroll-container">